requirements.txt
albumentations==1.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
efficientnet-pytorch==0.7.1
filelock==3.12.2
fonttools==4.40.0
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.1
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
munch==3.0.0
networkx==3.1
numpy==1.24.3
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.3
patchify==0.2.3
Pillow==9.5.0
pretrainedmodels==0.7.4
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.31.0
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
segmentation-models-pytorch==0.3.3
six==1.16.0
split-folders==0.5.1
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
torch==2.0.1+cu118
torchvision==0.15.2+cu118
torchaudio==2.0.2+cu118
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
zipp==3.15.0
# Additional dependencies for STEGO model
seaborn==0.12.2